Create 20240201 Floats to Percentages
Write a function to_percent that transforms a given list of float values, ranging between 0 and 1, into percentages. (0 = 0, 0.5 = 50, 1 = 100)
Inside the function, create a new list.
Use a for loop to iterate through the given list of floats, and for each one, append its percent value to the new list.
The function will return the new list.